backend/teacher/requirements.txt
fastapi
uvicorn
passlib
bcrypt
cryptography
PyMySQL
python-jose
python-multipart
sqlalchemy
starlette
mysql-connector
decouple